rename tmy to meteo_yr
